refactor: move dicthtml conversion code to convert package
